+++ kong/tracing/propagation.lua
@@ -528,7 +528,9 @@ local function set(conf_header_type, found_header_type, proxy_span, conf_default
     kong.log.debug("skipping propagation of noop span")
     return
   end
-  set_propagated(proxy_span)
+  if reuse then
+    set_propagated(proxy_span)
+  end
 
   local set_header = kong.service.request.set_header