+++ spec/03-plugins/34-zipkin/zipkin_spec.lua
@@ -1481,3 +1481,69 @@ for _, strategy in helpers.each_strategy() do
 
   end)
 end
+
+for _, strategy in helpers.each_strategy() do
+  describe("Integration tests with instrumentations enabled", function()
+    local proxy_client, zipkin_client, service
+
+    setup(function()
+      local bp = helpers.get_db_utils(strategy, { "services", "routes", "plugins" })
+
+      service = bp.services:insert {
+        name = string.lower("http-" .. utils.random_string()),
+      }
+
+      -- kong (http) mock upstream
+      bp.routes:insert({
+        name = string.lower("route-" .. utils.random_string()),
+        service = service,
+        hosts = { "http-route" },
+        preserve_host = true,
+      })
+
+      -- enable zipkin plugin globally, with sample_ratio = 0
+      bp.plugins:insert({
+        name = "zipkin",
+        config = {
+          sample_ratio = 0,
+          http_endpoint = fmt("http://%s:%d/api/v2/spans", ZIPKIN_HOST, ZIPKIN_PORT),
+          default_header_type = "b3-single",
+        }
+      })
+
+      helpers.start_kong({
+        database = strategy,
+        nginx_conf = "spec/fixtures/custom_nginx.template",
+        tracing_instrumentations = "all",
+        tracing_sampling_rate = 1,
+      })
+
+      proxy_client = helpers.proxy_client()
+      zipkin_client = helpers.http_client(ZIPKIN_HOST, ZIPKIN_PORT)
+    end)
+
+    teardown(function()
+      helpers.stop_kong()
+    end)
+
+    it("generates spans for regular requests", function()
+      local start_s = ngx.now()
+
+      local r = proxy_client:get("/", {
+        headers = {
+          ["x-b3-sampled"] = "1",
+          host  = "http-route",
+          ["zipkin-tags"] = "foo=bar; baz=qux"
+        },
+      })
+      assert.response(r).has.status(200)
+
+      local spans = wait_for_spans(zipkin_client, 3, service.name)
+      local request_span = assert(get_span("get", spans), "request span missing")
+      local proxy_span = assert(get_span("get (proxy)", spans), "proxy span missing")
+
+      -- common assertions for request_span and proxy_span
+      assert_span_invariants(request_span, proxy_span, 16 * 2, start_s, "kong")
+    end)
+  end)
+end